feat(step2) - Creation of a project to study inheritance and polymorphism.
Create a class called account. This class will have the properties (attributes) account_number and balance and the methods deposit (to add value to the balance), withdrawal (to withdraw value from the balance, as long as it is not negative) and view balance (to show the balance value). Then, create a class called special_account, which inherits the account class and adds the limit property, which is an extra balance added to the balance. In the special_account class, the withdrawal method should be rewritten to allow withdrawals, even if the balance is negative, up to the limit value.
Install an object of the special account type and assign the initial values ​​to this account to test the methods previously written in exercise 3, creating a menu that allows you to choose the withdrawal, deposit or exit options.